CNDB-15001: Split VectorLocalTest to prevent timeouts
